Produce appropriate build logs
The output from both Wine and DosBox is written to `src/BUILD.LOG`.
This consists of what Borland `make` and the other build tools it
invokes, such as `tasm` and `bcc32` would print to the terminal.

In order to capture the output from tools run inside DosBox, the
configuration file written by Wine is patched to add a redirect to a
file. This file is later written to stdout.